+++ src/app/core/utils/download-file.utils.spec.ts
@@ -1,0 +1,44 @@
+/**
+ * Copyright 2021 Vitagroup AG
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { downloadFile } from './download-file.utils'
+
+describe('Download file utils', () => {
+  beforeEach(() => {
+    jest.spyOn(document.body, 'appendChild')
+  })
+  afterEach(() => {
+    jest.clearAllMocks()
+  })
+  it('should create the link for csv export correctly', () => {
+    const expectedLink = document.createElement('a')
+    expectedLink.setAttribute('href', 'https://test-link.com')
+    expectedLink.setAttribute('download', 'csv_export_test.zip')
+    expectedLink.style.display = 'none'
+
+    downloadFile('test', 'csv', 'data')
+    expect(document.body.appendChild).toHaveBeenCalledWith(expectedLink)
+  })
+
+  it('should create the link for a json export correctly', () => {
+    const expectedLink = document.createElement('a')
+    expectedLink.setAttribute('href', 'data:text/json;charset=utf-8,test-data-string')
+    expectedLink.setAttribute('download', 'json_export_test.json')
+    expectedLink.style.display = 'none'
+    downloadFile('test', 'json', 'test-data-string')
+    expect(document.body.appendChild).toHaveBeenCalledWith(expectedLink)
+  })
+})