+++ src/app/modules/search/components/patient-filter/patient-filter.component.ts
@@ -13,26 +13,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { Component, OnInit } from '@angular/core'
+import { Component, OnDestroy, OnInit } from '@angular/core'
 import { Router } from '@angular/router'
-import { Observable } from 'rxjs'
+import { Observable, Subscription } from 'rxjs'
+import { filter, map, take } from 'rxjs/operators'
+import { CohortService } from 'src/app/core/services/cohort/cohort.service'
 import { PatientFilterService } from 'src/app/core/services/patient-filter/patient-filter.service'
+import { ProfileService } from 'src/app/core/services/profile/profile.service'
 import { ToastMessageService } from 'src/app/core/services/toast-message/toast-message.service'
 import { IDetermineHits } from 'src/app/shared/components/editor-determine-hits/determine-hits.interface'
+import { AvailableRoles } from 'src/app/shared/models/available-roles.enum'
 import { ICohortPreviewApi } from 'src/app/shared/models/cohort-preview.interface'
 import { ICohortGroupApi } from 'src/app/shared/models/project/cohort-group-api.interface'
 import { CohortGroupUiModel } from 'src/app/shared/models/project/cohort-group-ui.model'
 import { ProjectUiModel } from 'src/app/shared/models/project/project-ui.model'
 import { ToastMessageType } from 'src/app/shared/models/toast-message-type.enum'
+import { IUserProfile } from 'src/app/shared/models/user/user-profile.interface'
 
 @Component({
   selector: 'num-patient-filter',
   templateUrl: './patient-filter.component.html',
   styleUrls: ['./patient-filter.component.scss'],
 })
-export class PatientFilterComponent implements OnInit {
+export class PatientFilterComponent implements OnInit, OnDestroy {
+  private subscriptions = new Subscription()
+  userRoles: AvailableRoles[]
   determineHits: IDetermineHits = {
-    defaultMessage: 'AQL.HITS.MESSAGE_SET_ALL_PARAMETERS',
+    defaultMessage: 'QUERIES.HITS.MESSAGE_SET_ALL_PARAMETERS',
     count: null,
   }
   patientCount$: Observable<number>
@@ -46,7 +53,9 @@ export class PatientFilterComponent implements OnInit {
   constructor(
     private patientFilterService: PatientFilterService,
     private router: Router,
-    private toastMessageService: ToastMessageService
+    private toastMessageService: ToastMessageService,
+    private cohortService: CohortService,
+    private profileService: ProfileService
   ) {}
 
   ngOnInit(): void {
@@ -54,7 +63,23 @@ export class PatientFilterComponent implements OnInit {
     this.patientCount$ = this.patientFilterService.totalDatasetCountObservable$
     this.previewData$ = this.patientFilterService.previewDataObservable$
 
+    this.subscriptions.add(
+      this.profileService.userProfileObservable$
+        .pipe(
+          filter((profile) => profile.id !== undefined),
+          take(1),
+          map((profile: IUserProfile) => {
+            this.userRoles = profile.roles
+          })
+        )
+        .subscribe()
+    )
+
     this.patientFilterService.getAllDatasetCount().subscribe()
+  }
+
+  ngOnDestroy(): void {
+    this.subscriptions.unsubscribe()
   }
 
   setCurrentProject(): void {
@@ -84,10 +109,8 @@ export class PatientFilterComponent implements OnInit {
       try {
         this.updateDetermineHits(null, '', true)
         const cohortGroupApi: ICohortGroupApi = this.cohortNode.convertToApi()
-        const previewData = await this.patientFilterService
-          .getPreviewData(cohortGroupApi, false)
-          .toPromise()
-        this.updateDetermineHits(previewData.count, '')
+        const count = await this.getCount(cohortGroupApi).toPromise()
+        this.updateDetermineHits(count, '')
       } catch (error) {
         if (error.status === 451) {
           // *** Error 451 means too few hits ***
@@ -95,7 +118,19 @@ export class PatientFilterComponent implements OnInit {
         } else {
           this.updateDetermineHits(null, 'PROJECT.HITS.MESSAGE_ERROR_MESSAGE')
         }
+        // Reset the preview data to hide graphs and reset hits counter
+        this.patientFilterService.resetPreviewData()
       }
+    }
+  }
+
+  getCount(cohortGroupApi: ICohortGroupApi): Observable<number> {
+    if (this.userRoles.includes(AvailableRoles.Manager)) {
+      return this.patientFilterService
+        .getPreviewData(cohortGroupApi, false)
+        .pipe(map((value) => value.count))
+    } else {
+      return this.cohortService.getSize(cohortGroupApi, false)
     }
   }
 
@@ -107,7 +142,7 @@ export class PatientFilterComponent implements OnInit {
     } else {
       this.toastMessageService.openToast({
         type: ToastMessageType.Error,
-        message: 'PROJECT.NO_AQL_ERROR_MESSAGE',
+        message: 'PROJECT.NO_QUERY_ERROR_MESSAGE',
       })
     }
   }